requirements.txt
discord.py==2.3.2
pdfplumber==0.10.3
python-docx==1.1.0
pandas==2.2.2
openai==1.30.1
beautifulsoup4==4.12.3
requests==2.31.0
flask==3.0.3
